+++ packages/gateway/src/middleware/graphql-rate-limit/types.d.ts
@@ -1,0 +1,3 @@
+// these packages don't have correct @types, so I declare it to fix the error TS7016
+declare module 'graphql-cost-analysis/dist/costAnalysis';
+declare module 'graphql-depth-limit';