+++ packages/gateway/src/test/metric/metric-emitter.spec.ts
@@ -1,7 +1,6 @@
 import { EventEmitter } from 'events';
 import { subscribeToRequestMetrics } from '../../metric';
-import MetricsChannelsEnum from '../../metric/channels.enum';
-import PubSubEventsEnum from '../../metric/pubsub-events.enum';
+import { MetricsChannels, PubSubEvents } from '@graphql-portal/types';
 
 jest.mock('@graphql-portal/config', () => ({
   config: {
@@ -37,16 +36,16 @@ describe('hitRecord', () => {
 
     expect(emitter.on).toBeCalledTimes(6);
     expect(emitter.on).toBeCalledTimes(6);
-    expect(emitter.on).toHaveBeenNthCalledWith(1, MetricsChannelsEnum.GOT_REQUEST, expect.any(Function));
-    expect(emitter.on).toHaveBeenNthCalledWith(2, MetricsChannelsEnum.RESOLVER_CALLED, expect.any(Function));
-    expect(emitter.on).toHaveBeenNthCalledWith(3, MetricsChannelsEnum.RESOLVER_DONE, expect.any(Function));
-    expect(emitter.on).toHaveBeenNthCalledWith(4, MetricsChannelsEnum.RESOLVER_ERROR, expect.any(Function));
-    expect(emitter.on).toHaveBeenNthCalledWith(5, MetricsChannelsEnum.SENT_RESPONSE, expect.any(Function));
-    expect(emitter.on).toHaveBeenNthCalledWith(6, MetricsChannelsEnum.GOT_ERROR, expect.any(Function));
+    expect(emitter.on).toHaveBeenNthCalledWith(1, MetricsChannels.GOT_REQUEST, expect.any(Function));
+    expect(emitter.on).toHaveBeenNthCalledWith(2, MetricsChannels.RESOLVER_CALLED, expect.any(Function));
+    expect(emitter.on).toHaveBeenNthCalledWith(3, MetricsChannels.RESOLVER_DONE, expect.any(Function));
+    expect(emitter.on).toHaveBeenNthCalledWith(4, MetricsChannels.RESOLVER_ERROR, expect.any(Function));
+    expect(emitter.on).toHaveBeenNthCalledWith(5, MetricsChannels.SENT_RESPONSE, expect.any(Function));
+    expect(emitter.on).toHaveBeenNthCalledWith(6, MetricsChannels.GOT_ERROR, expect.any(Function));
 
     expect(subscribe).toBeCalledTimes(3);
-    expect(subscribe).toHaveBeenNthCalledWith(1, PubSubEventsEnum.RESOLVER_CALLED, expect.any(Function));
-    expect(subscribe).toHaveBeenNthCalledWith(2, PubSubEventsEnum.RESOLVER_DONE, expect.any(Function));
-    expect(subscribe).toHaveBeenNthCalledWith(3, PubSubEventsEnum.RESOLVER_ERROR, expect.any(Function));
+    expect(subscribe).toHaveBeenNthCalledWith(1, PubSubEvents.RESOLVER_CALLED, expect.any(Function));
+    expect(subscribe).toHaveBeenNthCalledWith(2, PubSubEvents.RESOLVER_DONE, expect.any(Function));
+    expect(subscribe).toHaveBeenNthCalledWith(3, PubSubEvents.RESOLVER_ERROR, expect.any(Function));
   });
 });