+++ packages/types/src/api-def-config.ts
@@ -9,6 +9,29 @@ export interface ApiDefConfig {
   enable_ip_filtering?: boolean;
   allow_ips?: string[];
   deny_ips?: string[];
+  /**
+   * Argument of the bytes package's method parse https://www.npmjs.com/package/bytes
+   */
+  request_size_limit?: string | number;
+  /**
+   * Maximum GraphQL query depth
+   */
+  depth_limit?: number;
+  /**
+   * Maximum complexity for a request
+   */
+  request_complexity_limit?: number;
+  /**
+   * Allowed queries summary complexity for a user in a period of time
+   */
+  rate_limit?: {
+    complexity: number;
+    per: number;
+  };
+  authentication?: {
+    auth_header_name?: string;
+    auth_tokens: string[];
+  };
   mesh?: {
     serve?: ServeConfig;
     require?: string[];
@@ -113,6 +136,8 @@ export interface Handler {
   soap?: SoapHandler;
   thrift?: ThriftHandler;
   tuql?: TuqlHandler;
+  ContentfulHandler?: ContentfulHandler;
+  SlackHandler?: SlackHandler;
   [k: string]: unknown;
 }
 export interface FhirHandler {
@@ -726,6 +751,22 @@ export interface TuqlHandler {
    * Path to the SQL Dump file if you want to build a in-memory database
    */
   infile?: string;
+}
+export interface ContentfulHandler {
+  /**
+   * Authentication token
+   */
+  token: string;
+  /**
+   * A endpoint of your Contentful GraphQL API
+   */
+  endpoint: string;
+}
+export interface SlackHandler {
+  /**
+   * Token, which can be obtained here: https://api.slack.com/authentication
+   */
+  token: string;
 }
 export interface Transform {
   /**